Correção na classe ClienteDAO - Delete

+++ Projeto/Computadores.com/src/java/com/computadores/dal/ClienteDAO.java
@@ -221,7 +221,7 @@ public class ClienteDAO implements IEntidadeDAO<Cliente> {
             // **/--/**
             //
 
-            pstmt.setLong(1, obj.getCodigo());
+            pstmt.setInt(1, obj.getCodigo());
 
             pstmt.execute();
         } catch (SQLException ex) {
